DocBotApp/backend/requirements.txt
fastapi
uvicorn
python-multipart
passlib[bcrypt]
bcrypt==4.0.1
PyJWT
python-docx
python-dotenv
